MUST VIEW
Before installing Solara, you must disable your antivirus. This is necessary to ensure that all game files are installed correctly and the application works correctly. If the antivirus is not disabled, false errors such as “Trojan” and others may occur. This happens because Solara is injected into Roblox and modifies or adds files. Enjoy watching everyone!